+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### changed 
+- A copy of the default grid is proposed if the default grid is not empty, and if the destination grid is not empty either.
 - It is no longer possible to delete files attached to the temporary version.
 
 ## Unreleased